pyproject.toml
[project]
name = "ai-baas"
version = "0.1.0"
description = "Universal backend for AI applications"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.31.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.14.0",
    "psycopg2-binary>=2.9.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "python-jose[cryptography]>=3.3.0",
    "bcrypt>=5.0.0",
    "python-multipart>=0.0.18",
    "email-validator>=2.2.0",
    "litellm>=1.55.0",
    "qdrant-client[fastembed]>=1.13.0",
    "PyMuPDF>=1.24.0",
    "python-docx>=1.1.0",
    "Pillow>=11.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "httpx>=0.27.0",
]
